// frontend/src/app/model/lekar.ts
import { Pregled } from "./pregled";

export class Lekar{
    firstname: string;
    lastname: string;
    username: string;
    password: string;
    type: number;
    phone:number;
    adress:string;
    photo: string;

    licence: number;
    specialization: string;
    office: string;
    email:string;
    pregledi: Pregled[]

}